refactor(Root): instalé dependencias globales y reubiqué la autenticación de usuario
